refactor: Extract run and static CLI commands from _main.py
Move the run command (with run_app, resolve_app, and friends) to
_main_run.py and the deprecated static/static-assets commands plus the
quarto glue (cells-to-app, get-shiny-deps) to _main_static.py, following
the _main_skills.py pattern. _main.py keeps the main group, command
registration, and the lazy-importing create/add-test wrappers, and
re-exports run_app for shiny.run_app.
